// apps/client/src/components/pdf/pdfhighlighter/types.ts
export type Page = {
  node: HTMLElement;
  /** 1-Index page number */
  number: number;
};